requirements.txt
requests==2.31.0
slack-sdk==3.27.1
pyyaml==6.0.1
python-dotenv==1.0.1 